refactor : domain, util 패키지 생성하여 클래스 특성에 따라 분리
