+++ src/main/java/step3/util/InputView.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.util;
 
 import java.util.Scanner;
