+++ src/main/java/step3/util/ResultView.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.util;
 
 public class ResultView {
