+++ src/main/java/step3/util/Validator.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.util;
 
 public class Validator {
