+++ src/test/java/study/step3/RacingTest.java
@@ -5,9 +5,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import step3.Car;
-import step3.Racing;
-import step3.Validator;
+import step3.domain.Car;
+import step3.domain.Racing;
+import step3.util.Validator;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
